+++ mockup/patterns/resourceregistry/js/overrides.js
@@ -14,8 +14,14 @@ define([
     className: 'tab-pane overrides',
     editing: null,
     canSave: false,
+    newFile: false,
 
     template: _.template(
+      '<div class="clearfix">' +
+        '<div class="btn-group pull-right">' +
+          '<button class="plone-btn plone-btn-primary add-file"><%- _t("Add file") %></button> ' +
+        '</div>' +
+      '</div>' +
       '<form class="row">' +
         '<div class="col-md-12">' +
           '<p><%- _t("Only ++plone++ resources are available to override") %></p>' +
@@ -23,7 +29,7 @@ define([
         '</div>' +
       '</form>' +
       '<div class="row">' +
-        '<div class="col-md-12">' +
+        '<div class="col-md-12 edit-area">' +
           '<% if(view.editing){ %>' +
             '<p class="resource-name text-primary"><%- view.editing %></p> ' +
             '<div class="plone-btn-group">' +
@@ -36,9 +42,10 @@ define([
         '</div>' +
       '</div>'),
     events: {
-      'click button.plone-btn-danger': 'itemDeleted',
-      'click button.plone-btn-primary': 'itemSaved',
-      'click button.plone-btn-default': 'itemCancel'
+      'click .edit-area button.plone-btn-danger': 'itemDeleted',
+      'click .edit-area button.plone-btn-primary': 'itemSaved',
+      'click .edit-area button.plone-btn-default': 'itemCancel',
+      'click .plone-btn.add-file': 'addFile'
     },
 
     initialize: function(options){
@@ -48,6 +55,20 @@ define([
 
     serializedModel: function(){
       return $.extend({}, { view: this }, this.options);
+    },
+
+    addFile: function(){
+      var resource = window.prompt('Enter full path and filename', '++plone++static/' + _t('you-filename.js'));
+      if(resource.indexOf('++plone++static/') === -1){
+        window.alert(_t('Filename must start with ++plone++static/'));
+      }else{
+        if(this.options.data.overrides.indexOf(resource) === -1){
+          this.options.data.overrides.push(resource);
+        }
+        this.newFile = true;
+        this.editing = resource;
+        this.render();
+      }
     },
 
     itemSaved: function(e){
@@ -144,30 +165,40 @@ define([
         if(url[url.length - 1] !== '/'){
           url += '/';
         }
-        that.tabView.loading.show();
-        $.ajax({
-          // cache busting url
-          url: url + that.editing + '?' + utils.generateId(),
-          dataType: 'text'
-        }).done(function(data){
-          var $pre = $('<pre class="pat-texteditor" />');
-          $pre.text(data);
-          that.$editorContainer.empty().append($pre);
-          that.editor = new TextEditor($pre, {
-            width: $('.editor').width(),
-            height: 500
+        if(that.newFile){
+          that.showEditor('');
+          that.newFile = false;
+        }else{
+          that.tabView.loading.show();
+          $.ajax({
+            // cache busting url
+            url: url + that.editing + '?' + utils.generateId(),
+            dataType: 'text'
+          }).done(function(data){
+            that.showEditor(data);
+          }).fail(function(){
+            window.alert(_t('error loading resource for editing'));
+            that.tabView.loading.hide();
           });
-          that.editor.setSyntax(that.editing);
-          that.tabView.loading.hide();
-          that.editor.editor.on('change', function(){
-            that.$el.find('.plone-btn-primary,.plone-btn-default').removeClass('disabled');
-          });
-        }).fail(function(){
-          window.alert(_t('error loading resource for editing'));
-          that.tabView.loading.hide();
-        });
+        }
       }
-    }
+    },
+
+    showEditor: function(data){
+      var that = this;
+      var $pre = $('<pre class="pat-texteditor" />');
+      $pre.text(data);
+      that.$editorContainer.empty().append($pre);
+      that.editor = new TextEditor($pre, {
+        width: $('.editor').width(),
+        height: 500
+      });
+      that.editor.setSyntax(that.editing);
+      that.tabView.loading.hide();
+      that.editor.editor.on('change', function(){
+        that.$el.find('.plone-btn-primary,.plone-btn-default').removeClass('disabled');
+      });
+    }    
   });
 
   return OverridesView;