+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.0.5 (unreleased)
 ------------------
 
+- capitalize "Save" buttons on resource registry
+  [vangheem]
+
 - Set pat-tooltip's html option to ``true`` by default, as it cannot be set by
   the options. Real fix has still to be done.
   [thet]